:sparkles: Completed Home Page
This probably will be the final version used for v1.0.0, so...
*confetti*

It seems that for several of the next pages I will need an image or two,
so I was thinking of either using a placeholder or dedicating some time
to actually look for the images. Probably should use placeholders.