:sparkles: Added SitesList selector
A really basic template for the sites selector
Again, lots of the front end and some styles are just there to show a
rough draft of how things are expected to look.
In the future many will be changed to use actual images or icons.
Right now I'm just concentrating on getting the functionality and the
base layout.