:dizzy: Navbar no longer uses redux state
I couldn't get rid of redux completely for this one since push still
needs it.

Obviously the big risk is not having redux as the whole source of truth,
but it's also true that for this components the rest of the app
shouldn't care about the them. They're stand-alone and they don't need
to save anything into the redux state, so why use it at all?

KISS