:dizzy: Moved auth constants out of redux state
It makes more sense for them to be an actual exportable constants in a
module or a config than for them to be dancing around the redux state
when they don't really change.

Just to make sure I double double checked that it's safe for the client
id to be on the front end. It is.